Use GitHub API to show gravatars of all contributors.
Normally I would cache this somewhere rather than calling on every page load, as the API is rate-limited to 60 requests per minute. However since the Twitter API is used without any kind of caching, I don't think it's a major issue - The Twitter API is rate limited in the same way as GitHub.