Fix broken download button on landing page
MetaCPAN API v0 has been deprecated and shut down, so we could never
fill in the download information for the button. The response is
slightly different in v1 than v0, so adjusted accordingly.

The version for Dancer2 doesn't fit as cleanly as the version for
Dancer, so I removed it (and version handling) entirely.